Move event bus implementation to its own package
