+++ tests/src/Shared/Infrastructure/Bus/Event/MySqlDoctrineDomainEventsConsumerTest.php
@@ -5,9 +5,9 @@ declare(strict_types = 1);
 namespace CodelyTv\Tests\Shared\Infrastructure\Bus\Event;
 
 use CodelyTv\Shared\Domain\Bus\Event\DomainEvent;
-use CodelyTv\Shared\Infrastructure\Bus\Event\MySqlDoctrineEventBus;
-use CodelyTv\Shared\Infrastructure\Bus\Event\MySqlDoctrineDomainEventsConsumer;
 use CodelyTv\Shared\Infrastructure\Bus\Event\DomainEventMapping;
+use CodelyTv\Shared\Infrastructure\Bus\Event\MySql\MySqlDoctrineDomainEventsConsumer;
+use CodelyTv\Shared\Infrastructure\Bus\Event\MySql\MySqlDoctrineEventBus;
 use CodelyTv\Tests\Mooc\Courses\Domain\CourseCreatedDomainEventMother;
 use CodelyTv\Tests\Mooc\CoursesCounter\Domain\CoursesCounterIncrementedDomainEventMother;
 use CodelyTv\Tests\Shared\Infrastructure\PhpUnit\InfrastructureTestCase;
