+++ tests/src/Shared/Infrastructure/Bus/Event/MySqlDoctrineEventBusTest.php
@@ -4,7 +4,7 @@ declare(strict_types = 1);
 
 namespace CodelyTv\Tests\Shared\Infrastructure\Bus\Event;
 
-use CodelyTv\Shared\Infrastructure\Bus\Event\MySqlDoctrineEventBus;
+use CodelyTv\Shared\Infrastructure\Bus\Event\MySql\MySqlDoctrineEventBus;
 use CodelyTv\Tests\Mooc\Courses\Domain\CourseCreatedDomainEventMother;
 use CodelyTv\Tests\Mooc\CoursesCounter\Domain\CoursesCounterIncrementedDomainEventMother;
 use CodelyTv\Tests\Shared\Infrastructure\PhpUnit\InfrastructureTestCase;
